divide tests to different python file and add more test for vote
